# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "data-loom.",
        "title": "data-loom API",
        "contact": {},
        "version": "v1.0"
    },
    "host": "localhost:8080",
    "basePath": "/data-loom/api/v1",
    "paths": {
...
        "/cluster/statistic/bu": {
            "get": {
                "description": "查询集群维度集群的大表统计信息",
                "tags": [
                    "统计信息"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "统计开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计大于等于该值的表",
                        "name": "start_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计小于等于该值的表",
                        "name": "end_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "BU",
                        "name": "bu",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ClusterStatistic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/cluster/statistic/cluster": {
            "get": {
                "description": "查询集群维度集群的大表统计信息",
                "tags": [
                    "统计信息"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "统计开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计大于等于该值的表",
                        "name": "start_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计小于等于该值的表",
                        "name": "end_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "BU",
                        "name": "bu",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "cluster_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ClusterStatistic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/cluster/statistic/database": {
            "get": {
                "description": "查询库维度集群的大表统计信息",
                "tags": [
                    "统计信息"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "统计开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计大于等于该值的表",
                        "name": "start_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计小于等于该值的表",
                        "name": "end_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "BU",
                        "name": "bu",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "cluster_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "库名",
                        "name": "database",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ClusterStatistic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/cluster/statistic/date": {
            "get": {
                "description": "查询集群维度集群的大表统计信息",
                "tags": [
                    "统计信息"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "统计开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计大于等于该值的表",
                        "name": "start_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计小于等于该值的表",
                        "name": "end_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ClusterStatistic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/cluster/statistic/table": {
            "get": {
                "description": "查询表维度集群的大表统计信息",
                "tags": [
                    "统计信息"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "统计开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计大于等于该值的表",
                        "name": "start_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "统计小于等于该值的表",
                        "name": "end_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "BU",
                        "name": "bu",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群ID",
                        "name": "cluster_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "cluster_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "库名",
                        "name": "database",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "table",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ClusterStatistic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "返回信息",
                    "type": "string"
                }
            }
        },
...
        "services.ClusterStatistic": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.ClusterStatisticDetail"
                    }
                },
                "end_date": {
                    "type": "string"
                },
                "end_size": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "start_size": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "services.ClusterStatisticDetail": {
            "type": "object",
            "properties": {
                "bu": {
                    "description": "bu",
                    "type": "string"
                },
                "cluster_id": {
                    "description": "集群ID",
                    "type": "string"
                },
                "cluster_name": {
                    "description": "集群名称",
                    "type": "string"
                },
                "database": {
                    "description": "库名",
                    "type": "string"
                },
                "date": {
                    "description": "统计的日期",
                    "type": "string"
                },
                "policies": {
                    "description": "对应的策略",
                    "type": "string"
                },
                "policies_num": {
                    "description": "策略个数",
                    "type": "integer"
                },
                "table": {
                    "description": "表名",
                    "type": "string"
                },
                "table_size": {
                    "description": "表大小(GB)",
                    "type": "integer"
                },
                "tables_num": {
                    "description": "大表个数",
                    "type": "integer"
                }
            }
        },
...
    }
}